requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-aws>=5.0.0,<6.0.0
jinja2>=3.0.0
pyyaml>=6.0